API integration for session timings and dates completed

+++ Formula1Tracker/include/Driver.h
@@ -23,7 +23,7 @@ public:
         return teamName;
     }
 
-    int getPoints() const{
+    float getPoints() const{
         return points;
     }
